Improve API mapper
